--- controllers/manualController.js
@@ -1,16 +1,0 @@
-const db = require("../models");
-
-//creating input in the DB
-
-module.exports = {
-    create: function (req, res) {
-        db.Manual.create(req.body)
-            .then(dbBook => res.json(dbBook))
-            .catch(err => res.status(422).json(err));
-    },
-    findById: function (req, res) {
-        db.Manual.findById(req.params.id)
-            .then(dbManual => res.json(dbManual))
-            .catch(err => res.status(422), json(err));
-    }
-};
