--- models/companyManual.js
@@ -1,16 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const companySchema = new Schema({
-    companyName: {type: String, required: true },
-    manual_data: [
-        {
-            type: Schema.Types.ObjectId,
-            ref: "Manual"
-        }
-    ]   
-})
-
-const Company = mongoose.model("Company", companySchema);
-
-module.exports = Company;
