+++ routes/api/manual.js
