+++ routes/api/old/info.js
@@ -1,4 +1,4 @@
-const db = require("../../models");
+const db = require("../../../models");
 const axios = require("axios");
